Modularizar leitura das cidades em uma função

+++ cidade.h
@@ -1,9 +1,12 @@
 #ifndef CIDADE_H
 #define CIDADE_H
 
+// Manter o tamanho máximo do id consistente mesmo em plataformas diferentes
+#include <stdint.h>
+
 typedef struct Cidade
 {
-    int id;
+    uint64_t id;
     double x;
     double y;
 } Cidade;
